requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-dotenv==1.0.1
requests==2.31.0
groq>=0.9.0
streamlit